Address review points

+++ x-pack/plugin/ml/src/main/java/org/elasticsearch/xpack/ml/dataframe/process/AnalyticsProcessManager.java
@@ -200,13 +200,16 @@ public class AnalyticsProcessManager {
         public synchronized void stop() {
             LOGGER.debug("[{}] Stopping process", id);
             processKilled = true;
-            dataExtractor.cancel();
-            try {
-                process.kill();
-            } catch (IOException e) {
-                LOGGER.error(new ParameterizedMessage("[{}] Failed to kill process", id), e);
+            if (dataExtractor != null) {
+                dataExtractor.cancel();
             }
-
+            if (process != null) {
+                try {
+                    process.kill();
+                } catch (IOException e) {
+                    LOGGER.error(new ParameterizedMessage("[{}] Failed to kill process", id), e);
+                }
+            }
         }
 
         /**
